Added some functionality to js and css
I fix the the animation so that it easy to be reversed, it is based on
a function that checks wether to see if there is an animation set.
Based on that every time the runTarget is clicked the function will be
acting like an switch between this animation. This way I dont have to
write that much css. So no need to add extra classes and keyframes to
simple reverse it.